requirements.txt
bcrypt==4.2.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
Werkzeug==3.0.4

